backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.0
Flask-Limiter==3.7.0
SQLAlchemy==2.0.31
pydantic==2.8.2
python-dateutil==2.9.0.post0
requests==2.32.3
python-evtx==0.7.4
PyYAML==6.0.2
gunicorn==22.0.0
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
psycopg2-binary==2.9.9
